Support and test redirecting. (#27)
Notably, iron-location should now support a redirection happening immediately
after it wakes up and reads the values from the URL.

We still do not support writing to an iron-location's properties before it has
read from the URL (e.g. in ready() or attached() of other elements). This is
intentional, as it's almost always a mistake to redirect without first
examining the URL of the initial page load, as that's a major signal of user
intent.